requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openai>=0.27.0
finnhub-python